+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/camera/CameraCapturerFactory.kt
@@ -27,8 +27,8 @@ class CameraCapturerFactory {
          * 複数のカメラがある場合はフロントのカメラを優先します.
          *
          * @param context application context
-         * @param fixedResolution true の場合は解像度維持を優先、false の場合は
-         * フレームレート維持を優先する. デフォルト値は false.
+         * @param fixedResolution このパラメータは非推奨です。代わりに SoraMediaOption.degradationPreference を使用してください。
+         * このパラメータは互換性のためのみ残されており、実際には使用されません。
          * @param frontFacingFirst true の場合はフロントカメラを優先、false の場合は
          * リアカメラを優先して選択する. デフォルト値は true.
          * @return 生成された `CameraVideoCapturer`
@@ -58,15 +58,7 @@ class CameraCapturerFactory {
             }
 
             SoraLogger.d(TAG, "camera capturer => $videoCapturer")
-            return when (videoCapturer.isScreencast) {
-                fixedResolution -> {
-                    videoCapturer
-                }
-                else -> {
-                    SoraLogger.d(TAG, "Wrap capturer: original.isScreencast=${videoCapturer.isScreencast}, fixedResolution=$fixedResolution")
-                    CameraVideoCapturerWrapper(videoCapturer, fixedResolution)
-                }
-            }
+            return videoCapturer
         }
 
         private fun createCapturer(enumerator: CameraEnumerator, frontFacingFirst: Boolean): CameraVideoCapturer? {